fix w3 bai 5 6 7 8 9

+++ w3/bai5.c
@@ -1,10 +1,10 @@
 #include <stdio.h>
 
 int main() {
-    float x, y, z, sum, aver;
-    scanf("%f%f%f", &x, &y, &z);
+    double x, y, z, sum, aver;
+    scanf("%lf%lf%lf", &x, &y, &z);
     sum = x + y + z;
     aver = sum / 3;
-    printf("%f\n%f\n", sum, aver);
+    printf("%lf\n%lf\n", sum, aver);
     return 0;
 }
