+++ w3/bai7.c
@@ -1,11 +1,11 @@
 #include <stdio.h>
 
 int main() {
-    float X, Y, A, B, C;
-    scanf("%f%f", &X, &Y);
+    double X, Y, A, B, C;
+    scanf("%lf%lf", &X, &Y);
     A = X * X + Y * Y;
     B = (X + Y) * (X + Y);
     C = (X - Y) * (X - Y);
-    printf("%f\n%f\n%f\n", A, B, C);
+    printf("%lf\n%lf\n%lf\n", A, B, C);
     return 0;
 }
