+++ w3/bai8.c
@@ -2,9 +2,9 @@
 #include <stdio.h>
 
 int main() {
-    float x, y, z, f;
-    scanf("%f%f%f", &x, &y, &z);
+    double x, y, z, f;
+    scanf("%lf%lf%lf", &x, &y, &z);
     f = (x + y + z) / (x * x + y * y + 1) - fabs(x - z * cos(y));
-    printf("%f\n", f);
+    printf("%lf\n", f);
     return 0;
 }
