+++ w3/bai9.c
@@ -2,12 +2,12 @@
 #include <stdio.h>
 
 int main() {
-    float x, y, f = 0;
+    double x, y, f = 0;
     int i, j;
-    scanf("%f%f", &x, &y);
+    scanf("%lf%lf", &x, &y);
     for (i = 0; i <= 6; i++) {
         f += pow(x, i) * pow(y, 6 - i);
     }
-    printf("%f\n", f);
+    printf("%lf\n", f);
     return 0;
 }
